refactor: extract test
