+++ src/declarations/_types.ts
@@ -8,10 +8,10 @@ import type * as OrbiterDid006 from '$declarations/deprecated/orbiter-0-0-6.did'
 import type * as OrbiterDid007 from '$declarations/deprecated/orbiter-0-0-7.did';
 import type * as OrbiterDid008 from '$declarations/deprecated/orbiter-0-0-8.did';
 import type * as OrbiterDid020 from '$declarations/deprecated/orbiter-0-2-0.did';
-import type * as SatelliteDid016 from '$declarations/deprecated/satellite-0-1-6.did';
 import type * as SatelliteDid0016 from '$declarations/deprecated/satellite-0-0-16.did';
 import type * as SatelliteDid0017 from '$declarations/deprecated/satellite-0-0-17.did';
 import type * as SatelliteDid008 from '$declarations/deprecated/satellite-0-0-8.did';
+import type * as SatelliteDid016 from '$declarations/deprecated/satellite-0-1-6.did';
 import type * as ICDid from '$declarations/ic-management/ic-management.did';
 import type * as MissionControlDid from '$declarations/mission_control/mission_control.did';
 import type * as ObservatoryDid from '$declarations/observatory/observatory.did';
@@ -38,6 +38,6 @@ export type {
 	SatelliteDid0016,
 	SatelliteDid0017,
 	SatelliteDid008,
-	SputnikDid,
-	SatelliteDid016
+	SatelliteDid016,
+	SputnikDid
 };
