+++ src/tests/specs/console/console.controllers.spec.ts
@@ -1,11 +1,11 @@
-import { type ConsoleActor, type ConsoleDid, idlFactoryConsole } from '$declarations';
+import { type ConsoleActor, idlFactoryConsole } from '$declarations';
 import { type Actor, PocketIc } from '@dfinity/pic';
-import { assertNonNullish } from '@dfinity/utils';
 import { AnonymousIdentity } from '@icp-sdk/core/agent';
 import { Ed25519KeyIdentity } from '@icp-sdk/core/identity';
 import { inject } from 'vitest';
 import { CONTROLLER_ERROR_MSG } from '../../constants/console-tests.constants';
 import { CONTROLLER_METADATA } from '../../constants/controller-tests.constants';
+import { testControllers } from '../../utils/controllers-tests.utils';
 import { CONSOLE_WASM_PATH } from '../../utils/setup-tests.utils';
 
 describe('Console > Controllers', () => {
@@ -88,50 +88,9 @@ describe('Console > Controllers', () => {
 			actor.setIdentity(controller);
 		});
 
-		it('should add and remove additional controller', async () => {
-			const { set_controllers, del_controllers, list_controllers } = actor;
-
-			const newController = Ed25519KeyIdentity.generate();
-
-			const controllerData: ConsoleDid.SetController = {
-				scope: { Admin: null },
-				expires_at: [123n],
-				kind: [{ Automation: null }],
-				metadata: [['hello', 'world']]
-			};
-
-			await set_controllers({
-				controllers: [newController.getPrincipal()],
-				controller: controllerData
-			});
-
-			const addControllers = await list_controllers();
-
-			expect(addControllers).toHaveLength(2);
-
-			expect(
-				addControllers.find(([p]) => p.toText() === controller.getPrincipal().toText())
-			).not.toBeUndefined();
-
-			const newAddedController = addControllers.find(
-				([p]) => p.toText() === newController.getPrincipal().toText()
-			);
-
-			assertNonNullish(newAddedController);
-
-			expect(newAddedController[1].metadata).toEqual(controllerData.metadata);
-			expect(newAddedController[1].scope).toEqual(controllerData.scope);
-			expect(newAddedController[1].expires_at).toEqual(controllerData.expires_at);
-			expect(newAddedController[1].kind).toEqual(controllerData.kind);
-
-			await del_controllers({
-				controllers: [newController.getPrincipal()]
-			});
-
-			const updatedControllers = await list_controllers();
-
-			expect(updatedControllers).toHaveLength(1);
-			expect(updatedControllers[0][0].toText()).toEqual(controller.getPrincipal().toText());
+		testControllers({
+			actor: () => actor,
+			controller: () => controller
 		});
 	});
 });
