+++ src/tests/specs/observatory/observatory.spec.ts
@@ -1,6 +1,6 @@
-import { idlFactoryObservatory, type ObservatoryActor, type SatelliteDid } from '$declarations';
+import { idlFactoryObservatory, type ObservatoryActor } from '$declarations';
 import { PocketIc, type Actor } from '@dfinity/pic';
-import { assertNonNullish, toNullable } from '@dfinity/utils';
+import { toNullable } from '@dfinity/utils';
 import { AnonymousIdentity } from '@icp-sdk/core/agent';
 import { Ed25519KeyIdentity } from '@icp-sdk/core/identity';
 import { Principal } from '@icp-sdk/core/principal';
@@ -11,6 +11,7 @@ import {
 	CALLER_NOT_ANONYMOUS_MSG,
 	CALLER_NOT_CONTROLLER_OBSERVATORY_MSG
 } from '../../constants/observatory-tests.constants';
+import { testControllers } from '../../utils/controllers-tests.utils';
 import { OBSERVATORY_WASM_PATH } from '../../utils/setup-tests.utils';
 
 describe('Observatory', () => {
@@ -173,50 +174,9 @@ describe('Observatory', () => {
 			actor.setIdentity(controller);
 		});
 
-		it('should add and remove additional controller', async () => {
-			const { set_controllers, del_controllers, list_controllers } = actor;
-
-			const newController = Ed25519KeyIdentity.generate();
-
-			const controllerData: SatelliteDid.SetController = {
-				scope: { Admin: null },
-				expires_at: [123n],
-				kind: [{ Automation: null }],
-				metadata: [['hello', 'world']]
-			};
-
-			await set_controllers({
-				controllers: [newController.getPrincipal()],
-				controller: controllerData
-			});
-
-			const addControllers = await list_controllers();
-
-			expect(addControllers).toHaveLength(2);
-
-			expect(
-				addControllers.find(([p]) => p.toText() === controller.getPrincipal().toText())
-			).not.toBeUndefined();
-
-			const newAddedController = addControllers.find(
-				([p]) => p.toText() === newController.getPrincipal().toText()
-			);
-
-			assertNonNullish(newAddedController);
-
-			expect(newAddedController[1].metadata).toEqual(controllerData.metadata);
-			expect(newAddedController[1].scope).toEqual(controllerData.scope);
-			expect(newAddedController[1].expires_at).toEqual(controllerData.expires_at);
-			expect(newAddedController[1].kind).toEqual(controllerData.kind);
-
-			await del_controllers({
-				controllers: [newController.getPrincipal()]
-			});
-
-			const updatedControllers = await list_controllers();
-
-			expect(updatedControllers).toHaveLength(1);
-			expect(updatedControllers[0][0].toText()).toEqual(controller.getPrincipal().toText());
+		testControllers({
+			actor: () => actor,
+			controller: () => controller
 		});
 	});
 });
